foxglynn stdout -> stderr
git-svn-id: https://svn.lsv.ens-cachan.fr/barbot/Cosmos@641 c602a718-a262-424a-925f-c628338d67b4

+++ trunk/SOURCES/ModelGenerator/parameters.hpp
@@ -26,7 +26,7 @@ struct parameters {
 	bool RareEvent;
 	bool DoubleIS;
 	int BoundedRE;
-    int horizon;
+    double horizon;
     bool BoundedContinuous;
     
 	bool GMLinput;
